Revert!: Remove package =common-lisp-snippets=.
With SLY's bulit-in company for autocompletion and =SPC k= in Spacemacs to
semantically edit lisp, it's arguable how useful this snippet package is. It
ought to be like Latex snippets, where it must necessarily be customized.

Looking over the snippets, they are all basic and opinionated. Since remembering
the mini-DSL's of LOOP and FORMAT are a hassle, snippets for those would be
nice. But the HyperSpec already covers such topics and this package has no such
snippets. Still, I could someday give the package a shot.

Functionality is also duplicated. Here is the repository for a collection of
yasnippet snippets for many more languages, and this package is already included
in the spacemacs =auto-completion= layer.

https://github.com/AndreaCrotti/yasnippet-snippets