Feat: Add func SPACEMACS/SLY-MREPL-DWIM and bind to ", '".
", s i" is used to start the interactive SLY REPL. ", '" is a good default and
calls SLY-MREPL-SYNC instead. This sets the REPL directory and package namespace
to the current file. See SLY-MREPL as well.

This makes working with ASDF package inferred systems easier. The form

  (if (sly-current-connection)
      (call-interactively 'sly-mrepl-sync)
      (progn
      (call-interactively 'sly)
      (call-interactively 'sly-mrepl-sync)))

Does not work. So the behaviour right now is to call SLY if no current SLY
connection can be found.